refactor: Refactoring ProblemDetail, Result directory
