+++ src/pages/Result/components/Chart.tsx
